Create LinearRegression_UMAPDim_LearnedMotifPAM.py
To identify motifs most strongly associated with the observed variation in the UMAP embedding, generated from the learned motif representations of D. melanogaster introns, we performed linear regression analysis. Using the statsmodels Python library (Seabold & Perktold, 2010), we modelled each UMAP dimension separately as the dependent variable, and PAM scores of the learned motifs were considered as the independent variables. Specifically, in the first model, the UMAP1 dimension was regressed on PAM scores of the learned motifs, whereas in the second model, the UMAP2 dimension was regressed on the PAM scores of the learned motifs. This approach allowed us to rank motifs by the magnitude of their regression weights and the t-stat values, highlighting those with the greatest contribution to the spread along UMAP1 and UMAP2, and thus explaining the patterns observed along two dimensions.